requirements.txt
python-dotenv
Flask
telethon
pyTelegramBotAPI